refactor: Mentee 레벨 관련 파일 분리
- mentee_level.dart: MenteeLevelInfo 모델 정의
- mentee_level_table.dart: 온도 구간별 색상 및 이미지 매핑 정의
- get_mentee_level.dart: 현재 온도로부터 MenteeLevelInfo 반환하는 유틸 함수 정의
