+++ lib/features/mypage/data/model/mentee_level.dart
@@ -1,0 +1,17 @@
+import 'dart:ui';
+
+class MenteeLevelInfo {
+  final int level;
+  final Color color;
+  final int minDegree;
+  final int maxDegree;
+  final String image;
+
+  const MenteeLevelInfo({
+    required this.level,
+    required this.color,
+    required this.minDegree,
+    required this.maxDegree,
+    required this.image,
+  });
+}
